+++ src/main/java/com/revolvingmadness/sculk/language/lexer/Token.java
@@ -54,19 +54,10 @@ public class Token {
     }
 
     public boolean isRelationOperator() {
-        return this.type == TokenType.GREATER_THAN || this.type == TokenType.GREATER_THAN_OR_EQUAL_TO || this.type == TokenType.LESS_THAN || this.type == TokenType.LESS_THAN_OR_EQUAL_TO || this.type == TokenType.INSTANCE_OF || this.type == TokenType.SPACESHIP;
+        return this.type == TokenType.GREATER_THAN || this.type == TokenType.GREATER_THAN_OR_EQUAL_TO || this.type == TokenType.LESS_THAN || this.type == TokenType.LESS_THAN_OR_EQUAL_TO || this.type == TokenType.INSTANCEOF || this.type == TokenType.SPACESHIP;
     }
 
     public boolean isUnaryOperator() {
         return this.type == TokenType.HYPHEN || this.type == TokenType.EXCLAMATION_MARK || this.type == TokenType.PLUS || this.type == TokenType.DOUBLE_PLUS || this.type == TokenType.DOUBLE_HYPHEN;
-    }
-
-    @Override
-    public String toString() {
-        if (this.value == null) {
-            return "Token(type=" + this.type + ")";
-        }
-
-        return "Token(type=" + this.type + ", value=" + this.value + ")";
     }
 }